third_party/WebKit/LayoutTests/fast/canvas/draw-focus-if-needed-scrolls-horiz-expected.txt

PASS window.pageXOffset is 0
PASS window.pageXOffset is >= minXOffset
PASS maxXOffset is >= window.pageXOffset
PASS window.pageXOffset is >= minXOffset
PASS maxXOffset is >= window.pageXOffset
PASS window.pageXOffset is >= minXOffset
PASS maxXOffset is >= window.pageXOffset
PASS successfullyParsed is true

TEST COMPLETE
1000-pixel box
